Obsluga nieistniejacego katalogu

+++ src/s3_dir_backup.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-
+EXIT=0
 trap "quit" SIGINT
 trap "quit" SIGQUIT
 
@@ -72,30 +72,39 @@ function pauza_cycle(){
 
 for DIR in `cat $DIR_LIST_FILE`;
 	do
-		log "Backup directory $DIR"
-		FILEDATE=$(date +'%Y-%m-%d_%H%M%S')
-		BACKUP_FILE=$TEMP_DIR/$(basename $DIR)-$FILEDATE.tar.gz
-		tar -czf $BACKUP_FILE $DIR
-		EXIT_ERR=$?
-		# 0 = OK
-		if [ $EXIT_ERR -eq 0 ]
+		if [ -d "$DIR" ]
 			then
-				s3cmd put --progress --recursive --reduced-redundancy $BACKUP_FILE s3://$BUCKET_NAME > /dev/null
-				S3ERR=$?
-				if [ $S3ERR -eq 0 ]
+				log "Backup directory $DIR"
+				FILEDATE=$(date +'%Y-%m-%d_%H%M%S')
+				BACKUP_FILE=$TEMP_DIR/$(basename $DIR)-$FILEDATE.tar.gz
+				tar -czf $BACKUP_FILE $DIR
+				EXIT_ERR=$?
+				# 0 = OK
+				if [ $EXIT_ERR -eq 0 ]
 					then
-						log "Backup $DIR: OK"
+						s3cmd put --progress --recursive --reduced-redundancy $BACKUP_FILE s3://$BUCKET_NAME > /dev/null
+						S3ERR=$?
+						if [ $S3ERR -eq 0 ]
+							then
+								log "Backup $DIR: OK"
+							else
+								log "Backup $DIR: s3cmd ERROR ($S3ERR)"
+								EXIT=1
+						fi
 					else
-						log "Backup $DIR: s3cmd ERROR ($S3ERR)"
+						log "Backup $DIR: ERROR ($EXIT_ERR)"
+						EXIT=1
 				fi
+				rm $BACKUP_FILE
+				while [ "$PAUZA" = "true" ]
+					do
+						log "Paused, waiting for SIGCONT to resume"
+   		 				sleep 10
+				done
 			else
-				log "Backup $DIR: ERROR ($EXIT_ERR)"
+				log "$DIR does not exist"
+				EXIT=1
 		fi
-		rm $BACKUP_FILE
-		while [ "$PAUZA" = "true" ]
-			do
-				log "Paused, waiting for SIGCONT to resume"
-    				sleep 10
-		done
 	done
 rm -rf $TEMP_DIR
+exit $EXIT
